requirements.txt
pydantic<2.0

yake
grpcio
grpcio-tools
requests
gradio==3.23.0


# NLP and AI Libraries
feedparser
diffusers
google-api-python-client
langdetect
googletrans==4.0.0rc1
transformers
scipy
emoji
# textblob
nltk

scikit-learn
networkx
matplotlib
numpy
python-dateutil
torch  
Pillow
